+++ deluge_web_client/__init__.py
@@ -1,11 +1,23 @@
 from deluge_web_client.client import DelugeWebClient
-from deluge_web_client.exceptions import DelugeWebClientError
+from deluge_web_client.exceptions import (
+    DelugeWebClientConnectionError,
+    DelugeWebClientDecodeError,
+    DelugeWebClientError,
+    DelugeWebClientHTTPError,
+    DelugeWebClientRPCError,
+    DelugeWebClientTimeoutError,
+)
 from deluge_web_client.schema import Response, TorrentOptions
 from deluge_web_client.state import TorrentState
 
 __all__ = (
     "DelugeWebClient",
+    "DelugeWebClientConnectionError",
+    "DelugeWebClientDecodeError",
     "DelugeWebClientError",
+    "DelugeWebClientHTTPError",
+    "DelugeWebClientRPCError",
+    "DelugeWebClientTimeoutError",
     "Response",
     "TorrentOptions",
     "TorrentState",